fix: 🐛 cells do not have properly initialized coordinates
Each virtual cell has a default (0, 0) coordinate. When using methods
like eraseScreen() it modifies all the cells, but misses the
coordinates, which remains (0, 0). It leads to unusual behaviour
sometimes and was fixed with this commit. Now, when Canvas initializes
an array of cells it sets its coordinates immediately based on index in
the array.